+++ client/store/index.js
@@ -3,8 +3,8 @@ import createLogger from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
-import classes from './classes'
-import dndClass from './dndClass'
+import classes from './dndApi/classes'
+import dndClass from './dndApi/dndClass'
 
 const reducer = combineReducers({user,classes,dndClass})
 const middleware = composeWithDevTools(
@@ -14,5 +14,5 @@ const store = createStore(reducer, middleware)
 
 export default store
 export * from './user'
-export * from './classes'
-export * from './dndClass'
+export * from './dndApi/classes'
+export * from './dndApi/dndClass'
